+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -35,7 +35,7 @@ public class EmployeServiceImpl implements IEmployeService {
 
 	public int ajouterEmploye(Employe employe) {
 		employeRepository.save(employe);
-		return employe.getId();
+		return (int) employe.getId();
 	}
 
 	public void mettreAjourEmailByEmployeId(String email, int employeId) {
